+++ desktop/build.gradle
@@ -1,4 +1,4 @@
-group 'pl.radoslawjaros'
+group 'pmstation'
 version '1.0-SNAPSHOT'
 
 apply plugin: 'java'
@@ -6,7 +6,7 @@ apply plugin: 'application'
 
 sourceCompatibility = 1.8
 
-mainClassName = "pl.radoslawjaros.Station"
+mainClassName = "pmstation.Station"
 
 repositories {
     mavenCentral()
